🧪 JGXD12 tests: split release apply suite
