+++ .agentplane/tasks/202605010645-JGXD12/README.md
@@ -1,10 +1,10 @@
 ---
 id: "202605010645-JGXD12"
 title: "AP-10: Split release apply tests"
-status: "TODO"
+status: "DOING"
 priority: "med"
 owner: "CODER"
-revision: 1
+revision: 4
 origin:
   system: "manual"
 depends_on:
@@ -14,19 +14,36 @@ tags:
 verify:
   - "bunx vitest run packages/agentplane/src/commands/release/apply*.test.ts && bun run test:release:critical"
 plan_approval:
-  state: "pending"
-  updated_at: null
-  updated_by: null
-  note: null
+  state: "approved"
+  updated_at: "2026-05-01T10:00:12.642Z"
+  updated_by: "ORCHESTRATOR"
+  note: "Approved after AP-09 closed on main; scope remains splitting release apply tests with shared testkit release fixtures."
 verification:
-  state: "pending"
-  updated_at: null
-  updated_by: null
-  note: null
-comments: []
-events: []
+  state: "ok"
+  updated_at: "2026-05-01T10:18:52.192Z"
+  updated_by: "CODER"
+  note: "Verified: release apply tests split into preflight, version mutation, apply flow, and push recovery suites; shared release helpers moved to @agentplane/testkit/release; oversized baseline ratcheted."
+commit: null
+comments:
+  -
+    author: "CODER"
+    body: "Start: split release apply tests into preflight, version mutation, apply flow, and push recovery suites with shared @agentplane/testkit/release fixtures."
+events:
+  -
+    type: "status"
+    at: "2026-05-01T10:00:44.868Z"
+    author: "CODER"
+    from: "TODO"
+    to: "DOING"
+    note: "Start: split release apply tests into preflight, version mutation, apply flow, and push recovery suites with shared @agentplane/testkit/release fixtures."
+  -
+    type: "verify"
+    at: "2026-05-01T10:18:52.192Z"
+    author: "CODER"
+    state: "ok"
+    note: "Verified: release apply tests split into preflight, version mutation, apply flow, and push recovery suites; shared release helpers moved to @agentplane/testkit/release; oversized baseline ratcheted."
 doc_version: 3
-doc_updated_at: "2026-05-01T06:45:12.857Z"
+doc_updated_at: "2026-05-01T10:18:52.205Z"
 doc_updated_by: "CODER"
 description: "Split release apply tests by preflight, version mutation, apply flow, and push recovery while reusing testkit release fixtures."
 sections:
@@ -47,11 +64,24 @@ sections:
     3. Compare the final result against the task summary and touched scope. Expected: remaining follow-up is either resolved or explicit in ## Findings.
   Verification: |-
     <!-- BEGIN VERIFICATION RESULTS -->
+    ### 2026-05-01T10:18:52.192Z — VERIFY — ok
+    
+    By: CODER
+    
+    Note: Verified: release apply tests split into preflight, version mutation, apply flow, and push recovery suites; shared release helpers moved to @agentplane/testkit/release; oversized baseline ratcheted.
+    
+    VerifyStepsRef: doc_version=3, doc_updated_at=2026-05-01T10:00:44.868Z, excerpt_hash=sha256:0d548dd436bdb03fc15e0d9ff31fc6a83de79ce496338151f75d231f7af3fb05
+    
     <!-- END VERIFICATION RESULTS -->
   Rollback Plan: |-
     - Revert task-related commit(s).
     - Re-run required checks to confirm rollback safety.
-  Findings: ""
+  Findings: |-
+    - Observation: Checks passed: bunx vitest run packages/agentplane/src/commands/release/apply*.test.ts --testTimeout 240000 --hookTimeout 240000 (4 files, 19 tests); bun run test:release:critical (4 files, 16 tests); node scripts/check-oversized-test-baseline.mjs --threshold-lines 1000; node scripts/check-vitest-projects.mjs; bunx vitest run packages/agentplane/src/cli/local-ci-selection.test.ts --testTimeout 60000 --hookTimeout 60000; bun run typecheck; bun run lint:core; bunx prettier --check touched files; git diff --check; bun run framework:dev:bootstrap; node .agentplane/policy/check-routing.mjs.
+      Impact: The release apply oversized monolith is replaced by four sub-1000-line suites and the oversized baseline drops to 13 entries / 14609 total lines.
+      Resolution: Ready for AP-10 commit and branch_pr PR flow.
+      Promotion: incident-candidate
+      Fixability: external
 id_source: "generated"
 ---
 ## Summary
@@ -80,6 +110,14 @@ Split release apply tests by preflight, version mutation, apply flow, and push r
 ## Verification
 
 <!-- BEGIN VERIFICATION RESULTS -->
+### 2026-05-01T10:18:52.192Z — VERIFY — ok
+
+By: CODER
+
+Note: Verified: release apply tests split into preflight, version mutation, apply flow, and push recovery suites; shared release helpers moved to @agentplane/testkit/release; oversized baseline ratcheted.
+
+VerifyStepsRef: doc_version=3, doc_updated_at=2026-05-01T10:00:44.868Z, excerpt_hash=sha256:0d548dd436bdb03fc15e0d9ff31fc6a83de79ce496338151f75d231f7af3fb05
+
 <!-- END VERIFICATION RESULTS -->
 
 ## Rollback Plan
@@ -88,3 +126,9 @@ Split release apply tests by preflight, version mutation, apply flow, and push r
 - Re-run required checks to confirm rollback safety.
 
 ## Findings
+
+- Observation: Checks passed: bunx vitest run packages/agentplane/src/commands/release/apply*.test.ts --testTimeout 240000 --hookTimeout 240000 (4 files, 19 tests); bun run test:release:critical (4 files, 16 tests); node scripts/check-oversized-test-baseline.mjs --threshold-lines 1000; node scripts/check-vitest-projects.mjs; bunx vitest run packages/agentplane/src/cli/local-ci-selection.test.ts --testTimeout 60000 --hookTimeout 60000; bun run typecheck; bun run lint:core; bunx prettier --check touched files; git diff --check; bun run framework:dev:bootstrap; node .agentplane/policy/check-routing.mjs.
+  Impact: The release apply oversized monolith is replaced by four sub-1000-line suites and the oversized baseline drops to 13 entries / 14609 total lines.
+  Resolution: Ready for AP-10 commit and branch_pr PR flow.
+  Promotion: incident-candidate
+  Fixability: external
