+++ .agentplane/tasks/202605010645-JGXD12/pr/github-title.txt
@@ -1,0 +1,1 @@
+AP-10: Split release apply tests (JGXD12)
