+++ .agentplane/tasks/202605010645-JGXD12/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605010645-JGXD12/release-apply-test-split",
+  "created_at": "2026-05-01T10:00:45.097Z",
+  "head_sha": "e0ac888d700232bf418fce03f13c8111d2266243",
+  "last_verified_at": "2026-05-01T10:18:52.192Z",
+  "last_verified_sha": "e0ac888d700232bf418fce03f13c8111d2266243",
+  "schema_version": 1,
+  "task_id": "202605010645-JGXD12",
+  "updated_at": "2026-05-01T10:00:45.097Z",
+  "verify": {
+    "status": "pass"
+  }
+}
