+++ packages/testkit/src/release.ts
@@ -1,4 +1,4 @@
-import { mkdir, mkdtemp, writeFile } from "node:fs/promises";
+import { mkdir, mkdtemp, readFile, readdir, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import path from "node:path";
 
@@ -20,6 +20,75 @@ export async function writeReleaseNotes(
   const releasesDir = path.join(root, "docs", "releases");
   await mkdir(releasesDir, { recursive: true });
   await writeFile(path.join(releasesDir, `v${version}.md`), body, "utf8");
+}
+
+export async function withDryRunReleaseMode<T>(work: () => Promise<T>): Promise<T> {
+  const originalDryRun = process.env.AGENTPLANE_RELEASE_DRY_RUN;
+  process.env.AGENTPLANE_RELEASE_DRY_RUN = "1";
+  try {
+    return await work();
+  } finally {
+    if (originalDryRun === undefined) {
+      delete process.env.AGENTPLANE_RELEASE_DRY_RUN;
+    } else {
+      process.env.AGENTPLANE_RELEASE_DRY_RUN = originalDryRun;
+    }
+  }
+}
+
+export async function listReleasePlanRuns(root: string): Promise<string[]> {
+  const dir = path.join(root, ".agentplane", ".release", "plan");
+  const runNames = await readdir(dir);
+  return runNames
+    .map((s) => s.trim())
+    .filter(Boolean)
+    .toSorted();
+}
+
+export async function writeWorkflowMode(root: string, mode: "direct" | "branch_pr"): Promise<void> {
+  const configPath = path.join(root, ".agentplane", "config.json");
+  const raw = JSON.parse(await readFile(configPath, "utf8")) as Record<string, unknown>;
+  raw.workflow_mode = mode;
+  await writeFile(configPath, `${JSON.stringify(raw, null, 2)}\n`, "utf8");
+}
+
+export async function writeReleasePushScripts(opts: {
+  root: string;
+  prepublishBody?: string;
+  prepublishFastBody?: string;
+  prepublishHeavyBody?: string;
+  registryBody: string;
+}): Promise<void> {
+  const heavyBody = opts.prepublishHeavyBody ?? opts.prepublishBody ?? "";
+  const fastBody =
+    opts.prepublishFastBody ??
+    `${String.raw`process.stdout.write('release prepublish fast ok\n');`}\n`;
+  await mkdir(path.join(opts.root, "scripts"), { recursive: true });
+  await writeFile(
+    path.join(opts.root, "package.json"),
+    `${JSON.stringify(
+      {
+        name: "release-test-root",
+        private: true,
+        scripts: {
+          "release:prepublish:fast": "node scripts/prepublish-fast.mjs",
+          "release:prepublish:heavy": "node scripts/prepublish-heavy.mjs",
+          "release:prepublish":
+            "node scripts/prepublish-fast.mjs && node scripts/prepublish-heavy.mjs",
+        },
+      },
+      null,
+      2,
+    )}\n`,
+    "utf8",
+  );
+  await writeFile(path.join(opts.root, "scripts", "prepublish-fast.mjs"), fastBody, "utf8");
+  await writeFile(path.join(opts.root, "scripts", "prepublish-heavy.mjs"), heavyBody, "utf8");
+  await writeFile(
+    path.join(opts.root, "scripts", "check-npm-version-availability.mjs"),
+    opts.registryBody,
+    "utf8",
+  );
 }
 
 export async function seedReleaseWorkspace(
