+++ scripts/lib/test-route-registry.mjs
@@ -497,7 +497,10 @@ const RELEASE_TEST_FILES = [
   "packages/agentplane/src/commands/release/release-check-script.test.ts",
   "packages/agentplane/src/commands/release/check-release-parity-script.test.ts",
   "packages/agentplane/src/commands/release/check-release-version-script.test.ts",
-  "packages/agentplane/src/commands/release/apply.test.ts",
+  "packages/agentplane/src/commands/release/apply.apply-flow.test.ts",
+  "packages/agentplane/src/commands/release/apply.preflight.test.ts",
+  "packages/agentplane/src/commands/release/apply.push-recovery.test.ts",
+  "packages/agentplane/src/commands/release/apply.version-mutation.test.ts",
   "packages/agentplane/src/cli/release-recovery-script.test.ts",
 ];
